To allow serialization of Lintian::CollScript, pass package names as strings instead of coderefs
The coderef prevents the class CollScript from being serialized. That
makes it much less convenient or even impossible to use it, for
example in IO::Async::Function with its multi process features.

This change makes the class CollScript more versatile. Stores the
module name as a string instead of a coderef.

One can still use the class inside a IO::Async::Function without this
change by "closing" on it, i.e.  just using it as a local variable,
but that also means that the function must be defined in the loop.

Also, uses '->can' from UNIVERSAL to check for the presence of the
instance method 'collect', which is cleaner and more elegant. That
idea was based on this exchange with mst on #perl-help:

11:37 <@mst> if ("Lintian::coll::$ppkg"->can('collect')) { would be nicer
11:38 <@mst> and in the child you can call "Lintian::coll::$ppkg"->collect(...)
11:39 <@mst> and then you get rid of a 'no strict' *and* the bug :)

Gbp-Dch: ignore