Simplify or replace safe_qx wherever the new usage will not work.
As part of the effort to drop IPC::Run, safe_qx will be reimplemented
using simpler components. The options hash is no longer available, and
pipes are not, either. Simplifies the usage of safe_qx so it is
suitable for the new implementation, or replace it when that is not
possible.

Gbp-Dch: ignore