Add collection helper functions based on IO::Async.
This will help with the gradual migration of the collection scripts to
IO::Async based process management. Provides the following functions:

    safe_qx_ioasync:

The existing helper function safe_qx has to be replaced gradually. It
may not be mixed with IO::Async (although this new function can). The
new function, on the other hand, cannot be mixed with older code that
still uses IPC::Run.

This provides a second function with a slightly different name for the
transition period.

The new function is located in lib/Lintian/Util.pm and not, like the
old one, in lib/Lintian/Command.pm. The plan is to get rid of the
latter.

    gzip:

Takes data from memory, such as an uncompressed file index, and stores
it in a gzip-compressed file.

    gunzip_file

Unzips a file and stores it in another.

    sort_file_index:

Sorts a file index in memory. Commonly used function in the collection
scripts.

Gbp-Dch: ignore