Undefine $IO::Async::Loop::ONE_TRUE_LOOP in Unpacker to allow nesting of IO::Async in collscripts.
Together with using IO::Async::Loop::Epoll, this change provided the
breakthrough to switching the collection scripts from IPC::Run to
IO::Async.

A big thank you to mst and LeoNerd on #io-async for helping to figure
this out. mst, who is a software consultant and available for hire,
identified the need to undefine the loop singleton. LeoNerd developed
IO::Async and recommended Epoll. Thank you to both!

12:53 < mst> so ... you could use a subloop via really_new ... or, yeah, undef($IO::Async::ONE_TRUE_LOOP)
[...]
12:56 < LeoNerd> Certainly, if you're trying to use a loop in both parent and child and -not- clearing that variable, then yes that would explain lots of
                 breakage

In summary, using the same singleton loop in nested usage is never a
good idea.  This change will likely become the default upstream soon,
but it may some time to propagate into a stable version of IO::Async.
Until such time, we have to specify it separately.

Gbp-Dch: ignore