Split out build routine as Test::Lintian::Build.
This was previously part of the t/bin/build-test-packages but could in
the future be invoked in multiple ways. The Gitlab CI setup, for
example, may build in a different order than a developer at home.

Splits the code to build a test package into a separate module.

Gbp-Dch: ignore