In tags, add arrow between correct and incorrect spelling.
With different forms of English usage around, it can take a moment to
realize which spelling is correct. This commit aids and reinforces the
mental process by adding an arrow between the correct and the
incorrect spelling.

The arrow appears in all corrections regardless of how many words are
involved. The arrow is unicode symbol U+2b9e.