In test runner, add option to accept all tag changes.
With many tests focused on a single tag, large scale changes appeared
unlikely when calibrating expected tags interactively. The individual
approval of tag changes is cumbersome even for twenty tests, however,
especially if the changes are reviewed again in git.

Adds an option 'a' to the interactive calibration which will accept
all subsequent tag changes.