Add script 'generate-qa-list.pl' to generate 'qa-list.txt' from UDD database.
The script sends all package information to stdout. Debugging
information, if enabled with '-d', will go to stderr.

Will process all of 'unstable' and 'experimental'. Aggregates the tags
from across all architectures and the two releases using a DISTINCT
query. While the resulting count may not correspond to the true number
on any particular architecture or release, it is probably most useful
for the maintainer.

Debug packages are curently not included. I am trying to find them in
UDD.

An automated invocation (in bash) would probably look similar to:

    TEMPFILE="$(mktemp)"
    "$LINTIAN_ROOT/reporting/generate-qa-list.pl" > "$TEMPFILE"
    mv "$TEMPFILE" "$HTML_DIR/qa-list.txt"

Gbp-Dch: ignore