Split tests internally into separate specifications for building and testing.
This large commit splits the layout for all tests into separate
instructions for building and for testing. In some ways, the new
layout is more cumbersome, but it has been waiting for about a year.
There are many technical considerations that made this step necessary.

The separation makes it possible to rebuild test packages adaptively
based on hash-based file manifests. it is expected to resolve
Bug#935907 in short order.

The separation also prepares the test suite for new and more complex
efforts to build test packages, for example in chroot environments.
That step may be necessary to resolve conflicting build dependencies
between test packages. For performance reasons, it may not be enabled
by default.

The separation furtermore improves the accuracy of Gitlab CI artifact
caching (Bug#930487). For example, commit 37a3d3ea triggered a rebuild
that was not needed. Only the expected tags had changed, but they were
in the same folder.

Finally, the separation is a prerequisite for the 'variations' branch,
which will permutate inputs to test how Lintian checks perform in more
real-life scenarios. A simple example would be to randomize version
numbers. A more meaningful example would be to test all source
formats.

The introduction of distinct file system trees with separate source
and build directories is very much less convenient than the simple
'run' directory from before. It is expected, however, that the new
structure will be extensible and reliable.

The decision was not made lightly. In some ways I regret its
necessity.

The old 'desc' file retained its name but was moved into a folder
called 'build-spec'. It is essentially a list of variables that are
used to fill the templates. The following values were transferred:

    Author
    Build-Command
    Build-Product
    Default-Build-Depends
    Description
    Dh-Compat-Level
    Distribution
    Extra-Build-Depends
    Package-Architecture
    Prev-Version
    Priority
    Section
    Skeleton
    Source
    Source-Format
    Testname
    Type
    Version

Other values were related only to how tests were selected and run, or
how Lintian was invoked. They were split into a new file 'test-option'
in a folder called 'test-spec'. The following values were transferred
there:

    Check
    Default-Lintian-Options
    Lintian-Command-Line
    Match-Strategy
    Options
    Output-Format
    Profile
    References
    Test-Against
    Test-Architectures
    Test-Conflicts
    Test-Depends
    Todo

An unexpected drawback was that the test name became part of the build
specification and not the test specification. That may change in the
future to accommodate common-sense expectations.

The two file names holding data for build and test specifications
should be kept different Their respective directories are often found
with commands like:

    find t/tags -name 'desc'

It is one of the few effective means to manipulate our large set of
test specifications accurately.